long->int32

+++ src/eg3dmap.c
@@ -97,8 +97,8 @@ struct TileObject *findNearestTileObject(uint32 worldX, uint32 worldY) {
                                 nearestTile.entry = g_curTileEntry;
                                 nearestTile.id = g;
                                 nearestTile.dist = q;
-                                nearestTile.x = worldX + (long)h;
-                                nearestTile.y = worldY + (long)j;
+                                nearestTile.x = worldX + (int32)h;
+                                nearestTile.y = worldY + (int32)j;
                             }
                         }
                     }
@@ -298,9 +298,9 @@ void drawModelPoint(int x, int y) {
 }
 
 // ==== seg000:0x378e ====
-void buildVertexSignMask(int screenX, int screenY) {
-    long bit;
-    int edgeIdx;
+void buildVertexSignMask(int16 screenX, int16 screenY) {
+    int32 bit;
+    int16 edgeIdx;
 
     bit = 1L;
     g_modelEdgeCount = (int16)(uint8)(*((*(char far **)&g_modelStreamPtr)++)) & 0x1f;
