+++ src/enbrief.h
@@ -6,7 +6,7 @@
 
 void animateFlightPath(int16 *gfxPage);
 void menuLabelsReset(void);
-long calcMissionScore(int16 param);
+int32 calcMissionScore(int16 param);
 void processMenuItems(MenuItem *items, int16 unused, int16 itemCount, int16 cursorStartX, int16 cursorStartY, int16 *gfxPage);
 int16 selectMenuItem(MenuItem *items, int16 unused, int16 itemCount, int16 *inputState, int16 *gfxPage);
 void computeMissionResult(void);
